Initial Upload
The game is mostly completed at this stage, but still has some bugs and other things to do. The player is a spaceship guarding a wall and must type the words belonging to the enemies in order to destroy them. Should an enemy collide with the player or breech the wall, the game ends.