Rework
This is almost a complete revamp of the file structure system that adds a lof of features. The most notable being the installer, which allows all the files to be directed to the approrpiate directories.
The core game is completed now, there are just the menu and file access systems to complete.